refactor(poller): decompose poll loops and lifecycle phases
Extract per-loop poller modules behind a shared createPoller timer helper
and split task-engine lifecycle into named, testable phase functions while
preserving poll cadences and startup ordering.
