+++ server/task-engine/lifecycle/types.ts
@@ -1,0 +1,8 @@
+export interface AddAgentOpts {
+  prompt?: string;
+  agent?: string | null;
+  label?: string;
+  model?: string | null;
+  skeleton?: string;
+  notify_agent_id?: string | null;
+}
